requirements.txt
pydantic>=2.0.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-dotenv>=0.19.0
mcp>=0.1.0
asyncio>=3.4.3 